+++ src/cucumber.spec.ts
@@ -15,15 +15,13 @@ function attach(this: { test: any }, body: any, details: any) {
 const fixture = new supportCodeLibrary.World({});
 const test = fixture.test;
 
+for (const beforeAllHook of supportCodeLibrary.beforeTestRunHookDefinitions) {
+    test.beforeAll(() => beforeAllHook.code.apply({}));
+}
+
 for (const feature of features) {
     const tests = feature.tests;
     test.describe(feature.feature as string, async () => {
-        for (const beforeAllHook of supportCodeLibrary.beforeTestRunHookDefinitions) {
-            test.beforeAll(async () => {
-                await beforeAllHook.code.apply({});
-            });
-        }
-
         const world = new supportCodeLibrary.World({
             log,
             attach,
@@ -39,28 +37,26 @@ for (const feature of features) {
                 for (const beforeHook of supportCodeLibrary.beforeTestCaseHookDefinitions) {
                     if (beforeHook.appliesToTestCase(testCase)) {
                         const hookName = beforeHook.name ?? 'Before';
-                        await test.step(hookName, async () => {
-                            await beforeHook.code.apply(world, [{
-                                pickle: testCase
-                            }]);
-                        });
+                        await test.step(hookName, () => beforeHook.code.apply(world, [{
+                            pickle: testCase
+                        }]));
                     }
                 }
                 for (const pickleStep of testCase.steps) {
                     await test.step(pickleStep.text, async () => {
                         for (const beforeStep of supportCodeLibrary.beforeTestStepHookDefinitions) {
                             if (beforeStep.appliesToTestCase(testCase)) {
-                                await beforeStep.code.apply(world, [{
+                                await test.step('Before Step', () => beforeStep.code.apply(world, [{
                                     pickle: testCase,
                                     pickleStep
-                                }]);
+                                }]));
                             }
                         }
                         const steps = supportCodeLibrary.stepDefinitions
                             .filter(stepDefinition => stepDefinition.matchesStepName(pickleStep.text));
                         if (steps.length === 0) throw new Error(`Step '${pickleStep.text}' is not defined`);
                         if (steps.length > 1) throw new Error(`'${pickleStep.text}' matches multiple step definitions`);
-                        const [step] = steps;
+                        const [ step ] = steps;
                         const { parameters} = await step.getInvocationParameters({
                             step: {
                                 text: pickleStep.text,
@@ -76,11 +72,11 @@ for (const feature of features) {
                         }
                         for (const afterStep of supportCodeLibrary.afterTestStepHookDefinitions) {
                             if (afterStep.appliesToTestCase(testCase)) {
-                                await afterStep.code.apply(world, [{
+                                await test.step('After Step', () => afterStep.code.apply(world, [{
                                     pickle: testCase,
                                     pickleStep,
                                     result
-                                }]);
+                                }]));
                             }
                         }
                         if (result.error) throw result.error;
@@ -89,19 +85,16 @@ for (const feature of features) {
                 for (const afterHook of supportCodeLibrary.afterTestCaseHookDefinitions) {
                     if (afterHook.appliesToTestCase(testCase)) {
                         const hookName = afterHook.name ?? 'After';
-                        await test.step(hookName, async () => {
-                            await afterHook.code.apply(world, [{
-                                pickle: testCase, result
-                            }]);
-                        });
+                        await test.step(hookName, () => afterHook.code.apply(world, [{
+                            pickle: testCase, result
+                        }]));
                     }
                 }
             })
         }
-        for (const afterAllHook of supportCodeLibrary.afterTestRunHookDefinitions) {
-            test.afterAll(async function () {
-                await afterAllHook.code.apply({});
-            });
-        }
     });
+}
+
+for (const afterAllHook of supportCodeLibrary.afterTestRunHookDefinitions) {
+    test.afterAll(() => afterAllHook.code.apply({}));
 }